src/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
boto3==1.19.7
botocore==1.22.7
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cramjam==2.4.0
fastparquet==0.7.1
Flask==2.0.2
flask-restx==0.5.1
fsspec==2021.10.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
jsonschema==4.1.2
MarkupSafe==2.0.1
numpy==1.21.3
pandas==1.3.4
psycopg2-binary==2.9.1
pyarrow==6.0.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
thrift==0.15.0
urllib3==1.26.7
Werkzeug==2.0.2
